refactor s3 data validation data.py
